Move info Popover to separate component
