+++ frontend/src/metabase/components/MetadataInfo/Popover/index.tsx
@@ -1,0 +1,1 @@
+export * from "./Popover";
