+++ frontend/src/metabase/components/MetadataInfo/TableInfoPopover/index.ts
@@ -1,1 +1,1 @@
-export { TableInfoPopover } from "./TableInfoPopover";
+export * from "./TableInfoPopover";
